projet final sans design

+++ app/src/main/java/com/example/lajan/Class/CarteBancaire.kt
@@ -2,7 +2,7 @@ package com.example.lajan.Class
 
 data class CarteBancaire(
     var idCb : Int = -1,
-    var numeroCarte : Int,
+    var numeroCarte : Long,
     var dateExpiration : Int,
     var typeCarte : String,
     var keyUserCarte : Int
